feat: update article preprocessing and add environment variables for Supabase
