+++ .github/workflows/deploy.yml
@@ -24,7 +24,10 @@ jobs:
         run: npm ci
 
       - name: Preprocess
-        run: npm run preprocess
+        run: npm run preprocess-articles
+        env:
+          SUPABASE_SECRET_KEY: ${{ secrets.SUPABASE_SECRET_KEY }}
+          PUBLIC_SUPABASE_URL: ${{ secrets.PUBLIC_SUPABASE_URL }}
 
       - name: Commit changes
         uses: EndBug/add-and-commit@v9
